Fixed bugs and some more functional fragments - Fixed some bugs which occurred while registering new users. - Fixed a bug with changing the profile picture - Changed the formations in this file. - Added balance Fragment to the side nav of signed in users. In it the current balance of the current user is displayed and refreshed in real time. - Added new function goHome. It can be called from anywhere with a fragment manager. Its purpose is to send the user back to his in app home screen.

+++ app/src/main/java/de/walhalla/app2/firebase/Firebase.java
@@ -118,7 +118,7 @@ public class Firebase {
                                             Log.e(TAG, "setAuth: ", e);
                                         }
                                         try {
-                                            p.setBalance(((Double) snapshot.get(Person.BALANCE)).floatValue());
+                                            p.setBalance(((Double) snapshot.get(Person.BALANCE)).longValue());
                                         } catch (Exception e) {
                                             CRASHLYTICS.recordException(e);
                                             Log.e(TAG, "setAuth: ", e);
